+++ examples/encode_recode_decode_simple.py
@@ -66,7 +66,7 @@ def main():
 
     # Assign the data buffer to the encoder so that we may start
     # to produce encoded symbols from it
-    encoder.set_symbols(data_in)
+    encoder.set_const_symbols(data_in)
 
     while not decoder2.is_complete():