+++ examples/perpetual.py
@@ -120,7 +120,7 @@ def main():
 
     # Assign the data buffer to the encoder so that we can
     # produce encoded symbols
-    encoder.set_symbols(data_in)
+    encoder.set_const_symbols(data_in)
 
     while not decoder.is_complete():
         # Encode a packet into the payload buffer