+++ examples/switch_systematic_on_off.py
@@ -51,7 +51,7 @@ def main():
 
     # Assign the data buffer to the encoder so that we may start
     # to produce encoded symbols from it
-    encoder.set_symbols(data_in)
+    encoder.set_const_symbols(data_in)
 
     print("Starting encoding / decoding")